backend/requirements.txt
Celery
Django
django-celery-beat 
django-celery-results # currently requires Django 4.0 or older
python-decouple
redis
psycopg2-binary
channels[daphne] 
channels_redis
djangorestframework
djangorestframework-simplejwt
django-allauth
dj-rest-auth
django-cors-headers
django-cleanup
flower
# aniref
deepdanbooru-onnx==0.0.8
gdown==4.6.4
gradio==4.1.1
scikit-image==0.20.0
torch
torchvision
ultralytics==8.0.54
transformers
opencv-python-headless